Update Java to jdk-23, package name renamed to is.hi.verzla
