--- a1/src/main/java/is/hi/hbv501g/a1/public/items.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<!--
-  Names: Ingólfur and Logi
-  Date: 17.09.24
-  Course: HBV501G Hugbúnaðarverkefni 1
--->
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>BestDeal</title>
-    <link rel="stylesheet" href="styles.css">
-    <link href="https://fonts.googleapis.com/css2?family=Share+Tech+Mono&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
-  </head>
-  <body>
-    <div id="navbar">
-      <nav id="user-navbar">
-        <a class="navbar-item" href="main.html">Home</a>
-        <a class="navbar-item" href="profile.html">My Profile</a>
-        <a class="navbar-item" href="checkout.html">Cart</a>
-      </nav>
-      <br>
-      <nav id="category-navbar">
-        <a class="navbar-item" href="items.html?category=clothing">Clothing</a>
-        <a class="navbar-item" href="items.html?category=kitchen">Kitchen Goods</a>
-        <a class="navbar-item" href="items.html?category=electronics">Electronics</a>
-      </nav>
-    </div>
-    <header id="item-page-header">
-      <h1 id="item-page-title">Shop -add js function- </h1>
-    </header>
-    <main>
-      <section id="item-page">
-        <div id="all-item-container">
-        <!-- Items will be dynamically loaded here -->
-        </div>
-      </section>
-    </main>
-    <footer>
-      <p>
-        © 2024 Verzla ehf. Allur réttur áskilinn.
-      </p>
-    </footer>
-    <script src="index.js"></script>
-  </body>
-</html>
