--- a1/src/main/java/is/hi/hbv501g/a1/public/login.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<!--
-  Names: Ingólfur and Logi
-  Date: 17.09.24
-  Course: HBV501G Hugbúnaðarverkefni 1
--->
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Login</title>
-  <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-  <div class="login-container">
-    <h1>Login</h1>
-    <form id="login-form">
-      <label for="username">Username:</label>
-      <input type="text" id="username" name="username" required>
-      <label for="password">Password:</label>
-      <input type="password" id="password" name="password" required>
-      <button type="submit" id="login-button">Login</button>
-    </form>
-  </div>
-  <footer>
-    <p>
-      © 2024 Verzla ehf. Allur réttur áskilinn.
-    </p>
-  </footer>
-  <script src="index.js"></script>
-</body>
-</html>
