--- a1/src/main/java/is/hi/hbv501g/a1/public/profile.html
@@ -1,129 +1,0 @@
-<!DOCTYPE html>
-<!--
-  Names: Ingólfur and Logi
-  Date: 17.09.24
-  Course: HBV501G Hugbúnaðarverkefni 1
--->
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>SHOP</title>
-    <link rel="stylesheet" href="styles.css">
-    <link rel="stylesheet" href="creative.css">
-    <link href="https://fonts.googleapis.com/css2?family=Share+Tech+Mono&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
-  </head>
-  <body>
-    <div id="navbar">
-      <nav id="user-navbar">
-        <a class="navbar-item" href="main.html">Home</a>
-        <a class="navbar-item" href="profile.html">My Profile</a>
-        <a class="navbar-item" href="checkout.html">Cart</a>
-      </nav>
-      <br>
-      <nav id="category-navbar">
-        <a class="navbar-item" href="items.html?category=clothing">Clothing</a>
-        <a class="navbar-item" href="items.html?category=kitchen">Kitchen Goods</a>
-        <a class="navbar-item" href="items.html?category=electronics">Electronics</a>
-      </nav>
-    </div>
-    <div class="profile-container">
-      <h1>Welcome, <span id="username">John Doe</span>!</h1>
-      <div class="profile-boxes">
-        <img src="img/duck.png" alt="Profile Image" class="profile-image">
-        <div class="profile-info">
-          <p><strong>Name:</strong> <span id="name">John Doe</span></p>
-          <p><strong>Email:</strong> <span id="email">john.doe@example.com</span></p>
-          <p><strong>Address:</strong> <span id="address">1234 Street, City, Country</span></p>
-          <p><strong>Phone:</strong> <span id="phone">(123) 456-7890</span></p>
-          <button id="view-sold-products">View Sold Products</button>
-          <button id="view-bought-products">View Bought Products</button>
-          <button id="logout-button">Logout</button>
-        </div>
-      </div>
-    </div>
-    <div id="sold-products-view" class="product-view">
-      <ul>
-        <li>
-          <div id="all-item-container">
-            <article class="item-container">
-              <a class="item" href="one-item.html">
-                <img src="img/background.png" alt="the item">
-                <div class="item-detail">
-                  <h3 class="item-name">item name</h3>
-                  <p class="item-price">$price</p>
-                </div>
-              </a>
-            </article>
-            <article class="item-container">
-              <a class="item" href="one-item.html">
-                <img src="img/background.png" alt="the item">
-                <div class="item-detail">
-                  <h3 class="item-name">item name</h3>
-                  <p class="item-price">$price</p>
-                </div>
-              </a>
-            </article>
-          </div>
-        </li>
-        <!-- Repeat for each sold product -->
-      </ul>
-    </div>
-    <div id="bought-products-view" class="product-view">
-      <ul>
-        <!-- Repeat for each bought product -->
-      </ul>
-    </div>
-    <div class="my-listings-container">
-      <h1>Your Listings</h1>
-      <div class="listings" id="listings-scroll">
-        <!-- Listings will be dynamically inserted here -->
-      </div>
-      <button id="open-modal-button">Sell Your Product</button>
-    </div>
-
-    <!-- Modal for adding new product -->
-    <div id="product-modal" class="modal">
-      <div class="modal-content">
-        <span class="close">&times;</span>
-        <form id="product-form">
-          <input type="text" id="product-name" placeholder="Product Name" required>
-          <input type="number" id="product-price" placeholder="Price" required>
-          <textarea id="product-description" placeholder="Description" required></textarea>
-          <input type="text" id="product-size" placeholder="Size" required>
-          <input type="file" id="product-image" accept="image/*" required>
-          <button type="button" id="add-product-button">Add Product</button>
-        </form>
-      </div>
-    </div>
-    <script src="index.js"></script>
-    <script>
-      document.addEventListener('DOMContentLoaded', async () => {
-        const userId = localStorage.getItem('userId');
-        if (!userId) {
-          window.location.href = 'login.html';
-        } else {
-          // Fetch user profile data and populate the profile fields
-          try {
-            const response = await fetch(`/users/${userId}`);
-            if (!response.ok) {
-              throw new Error('Failed to load user profile');
-            }
-            const user = await response.json();
-            document.getElementById('username').textContent = user.username;
-            document.getElementById('name').textContent = user.username;
-            document.getElementById('email').textContent = user.email;
-            // Add other profile fields if necessary
-          } catch (error) {
-            console.error('Error loading user profile:', error);
-          }
-        }
-      });
-    </script>
-    <footer></footer>
-      <p>
-        © 2024 Verzla ehf. Allur réttur áskilinn.
-      </p>
-    </footer>
-  </body>
-</html>
