+++ verzla/src/main/java/is/hi/verzla/VerzlaApplication.java
@@ -1,0 +1,13 @@
+package is.hi.verzla;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class VerzlaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(VerzlaApplication.class, args);
+	}
+
+}
