+++ verzla/src/test/java/is/hi/verzla/VerzlaApplicationTests.java
@@ -1,14 +1,13 @@
-package is.hi.hbv501g.a1;
+package is.hi.verzla;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class A1ApplicationTests {
+class VerzlaApplicationTests {
 
 	@Test
 	void contextLoads() {
 	}
 
 }
-//this team member is Ingólfur Bjarni Elíasson
